% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_normalization.R
\name{DSP_normalization}
\alias{DSP_normalization}
\title{nDSPA}
\usage{
DSP_normalization(df_probe, mat_val_all, method = "geomean", probes = "all")
}
\arguments{
\item{df_probe}{numeric}

\item{mat_val_all}{numeric}

\item{method}{Possible methods of normalization include "geomean" and "mean".}

\item{probes}{The probes argument checks the presence of all probes within controls.}
}
\value{
norm_mat numeric. This function returns a normalized matrix.
}
\description{
This function performs Data Normalization steps
}
