% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_QC.R
\name{DSP_QC}
\alias{DSP_QC}
\title{nDSPA}
\usage{
DSP_QC(anno, val_all_df, scalefactor, thresh_filt = FALSE, PCF_filt = FALSE)
}
\arguments{
\item{anno}{numeric. This is a dataframe.}

\item{val_all_df}{numeric. This is a dataframe}

\item{scalefactor}{numeric.}

\item{thresh_filt}{logical. Possible values are TRUE or FALSE.}

\item{PCF_filt}{logical. Possible values are TRUE or FALSE.}
}
\description{
This function performs Quality Control steps
}
