% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_Scale.R
\name{DSP_Scale}
\alias{DSP_Scale}
\title{nDSPA}
\usage{
DSP_Scale(df_anno, mat_val_all, method = "area")
}
\arguments{
\item{mat_val_all}{numeric. This is a data matrix.}

\item{method}{Possible values are "area" and "nuclei"}

\item{df_annonumeric.}{This is a data frame}
}
\value{
scaled_df numeric. This is a dataframe.
}
\description{
This function performs Scaling steps
}
