% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_SNR.R
\name{DSP_SNR}
\alias{DSP_SNR}
\title{nDSPA}
\usage{
DSP_SNR(df_probe, mat_val_all, method = "geomean", probes = "all")
}
\arguments{
\item{df_probe}{numeric. This is a data frame.}

\item{mat_val_all}{numeric. This is a data matrix}

\item{method}{Possible values are "mean" and "geomean". Default Value: "geomean"}

\item{probes}{Default value is "all"}
}
\value{
norm_mat numeric. This is a data matrix.
}
\description{
This function is DSP_SNR.
}
