% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspNormalization.R
\name{dspNormalization}
\alias{dspNormalization}
\title{nDSPA}
\usage{
dspNormalization(
  seobj,
  use = "gmean",
  probe.set = "all",
  use.assay = "erccScaled"
)
}
\arguments{
\item{seobj}{Provide ndspaExperiment object.}

\item{use}{Statistical method to be used for scaling.
Possible values are "mean" or "gmean".}

\item{probe.set}{Set of Probes to be used for Normalization.
Accepts a vector of custom Probes.
Default: "all" uses all probes tagged as "Control" in the rowData.}

\item{use.assay}{Define on which assay the normalization must be performed.
Use assays(seobj) or assayNames(seobj) to view assays saved in ndspaExperiment
object. Default:"erccScaled"}
}
\value{
norm_mat numeric. This function returns a normalized matrix.
}
\description{
This function performs Data Normalization steps
}
\examples{
\dontrun{
test <- dspNormalization(test, use = "gmean", probe.set="all",
use.assay = "erccScaled")
# OR
test <- dspNormalization(test, use = "gmean", probe.set="all",
use.assay = "area.scaled.data")
}
}
