% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_QC.R
\name{ERCC_Scale_factor}
\alias{ERCC_Scale_factor}
\title{{FOV=75, BD=c(0.1,2.25), SF=c(0.3,3), Min_nuc = 200, Min_Area = 16000}}
\usage{
ERCC_Scale_factor(df_probe, df_val_all)
}
\description{
{FOV=75, BD=c(0.1,2.25), SF=c(0.3,3), Min_nuc = 200, Min_Area = 16000}
}
