% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erccScaleFactor.R
\name{erccScaleFactor}
\alias{erccScaleFactor}
\title{nDSPA}
\usage{
erccScaleFactor(seobj)
}
\arguments{
\item{seobj}{S4Vector Provide ndspaExperiment object.}
}
\value{
seobj ndspaExperiment Object with ColData populated with ERCC ScaleFactors.
}
\description{
Enables calculation of ERCC ScaleFactors.
}
\examples{
\dontrun{
test <- erccScaleFactor(test)
}
}
