% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_IDs_PCF.R
\name{filter_IDs_PCF}
\alias{filter_IDs_PCF}
\title{nDSPA}
\usage{
filter_IDs_PCF(anno, scalefactor, PCF_min = 0.3, PCF_max = 3)
}
\arguments{
\item{anno}{numeric. This is a dataframe.}

\item{scalefactor}{numeric. This is data frame.}

\item{PCF_min}{numeric. Default Value: 0.3}

\item{PCF_max}{numeric. Default value: 3.0}
}
\value{
filtered_scalefactor numeric. This is a data frame.
}
\description{
This function filter IDs by positive control Scaling
}
