% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDSPA.R
\docType{package}
\name{nDSPA}
\alias{nDSPA}
\title{nDSPA: A package for quality metrics, preprocessing, visualization, and differential testing analysis of nanostring DSP & spatial omics data}
\description{
nDSPA provides useful functionality to:
-Import nanostring DSP data
-Easily QC, Filter, Scale and Normalize experimental data
-Quickly visualize spatial omics data
-Easily build models and perform statistical analysis on categorical spatial omics data
}
\section{nDSPA functions}{

...
}

