% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndspaExperiment-class.R
\docType{class}
\name{ndspaExperiment-class}
\alias{ndspaExperiment-class}
\alias{.ndspaEx}
\title{nDSPA}
\arguments{
\item{counts}{character. Specify the file name in .tsv, .csv or .xlsx}

\item{...}{Other arguments passed while creating Summarized/ SingleCell Experiment object.
Specify the group file with x,y coordinates (in .csv, .tsv and .xlsx format).}
}
\value{
ne Returns a ndspaExperiment object.
}
\description{
Enables easy loading of the data matrices provided by Nanostring.
}
