% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndspaInteractivePlots.R
\name{ndspaInteractivePlots}
\alias{ndspaInteractivePlots}
\title{nDSPA}
\usage{
ndspaInteractivePlots(seobj, use.assay = "counts")
}
\arguments{
\item{seobj}{S4Vector Provide ndspaExperiment object.}

\item{use.assay}{character Provide the assay name to be used for QC plots. Default "counts".}
}
\value{
A Shiny app object is returned for interactive data exploration of ndspaExperiment object.
}
\description{
Enables visualisation of QC Plots
}
\details{
The ndspaInteractivePlots function is let user generate QC plots to guage the data quality.
}
\examples{
\dontrun{
# To plot all QC graphs
ndspaInteractivePlots(test)

# To plot all QC graphs using erccScaled assay
ndspaInteractivePlots(test,use.assay ="erccScaled")

}
}
