% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndspaQC.R
\name{ndspaQC}
\alias{ndspaQC}
\title{nDSPA}
\usage{
ndspaQC(seobj, thresh_filt = NULL, PCF_filt = NULL, tag.only = FALSE)
}
\arguments{
\item{seobj}{S4Vector. Input ndspaExperiment object.}

\item{thresh_filt}{vector. Provide Segment displayed name IDs as a vector.}

\item{PCF_filt}{vector. Provide Segment displayed name IDs as a vector.}

\item{tag.only}{logical. To tag the cells that passes default QC criteria. Possible values are TRUE or FALSE.}
}
\value{
seobj Updated ndspaExperiment object with erccScaled assay added.
}
\description{
This function performs Quality Control steps.
}
\examples{
\dontrun{
# Run without filtering
test <- ndspaQC(test)

# Run with PCF Filter ids
test <- ndspaQC(test,PCF_filt=ids)

# Run in Tag only mode
test <- ndspaQC(test,PCF_filt=ids,tag.only=TRUE)

}
}
