% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndspaReducedims.R
\name{ndspaReducedims}
\alias{ndspaReducedims}
\title{nDSPA}
\usage{
ndspaReducedims(seobj, use.assay = "counts", type = "PCA", ...)
}
\arguments{
\item{seobj}{S4Vector Provide ndspaExperiment object.}

\item{use.assay}{character Provide the assay name to be used for QC plots. Default "counts".}

\item{type}{character Type of Dimensional Reduction method to be used ("PCA","tsne","umap"). Default: "PCA"}

\item{...}{arguments passed to or from prcomp or Rtsne depending on type of method opted for dimension reduction.}
}
\value{
Matrix containing the new representations for the ndspaExperiment object.
}
\description{
Adding low-dimensional representations to nDSPA object
}
\details{
The ndspaReducedims function let user perform Dimension Reduction given ndspaExperiment
object and returns a matrix with computed PCs or other low dimensional embedding depending on type of method used.
}
\examples{
\dontrun{

# Calculating PCA
test.pca <- ndspaReducedims(test)

# Calculating tSNE
test.tsne <- ndspaReducedims(test, type = "tsne")

#Calculating UMAP

test.umap <- ndspaReducedims(test, type = "umap")
# Saving results
reducedDims(test) <- list(PCA = test.pca$data, TSNE = test.tsne$data, UMAP=test.umap$data)
}

}
