% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndspaScale.R
\name{ndspaScale}
\alias{ndspaScale}
\title{nDSPA}
\usage{
ndspaScale(seobj, use = c("mean", "gmean"), method = "area")
}
\arguments{
\item{seobj}{Provide ndspaExperiment object.}

\item{use}{Statistical method to be used for scaling. Possible values are "mean" or "gmean".}

\item{method}{Variable used to scale data. Possible values are "area" and "nuclei". Default: "area".}
}
\value{
seobj Returns ndspaExperiment updated with scaled assay.
}
\description{
This function performs Scaling.
}
\examples{
\dontrun{
test <- ndspaScale(test,use ="gmean", method = "area")

# OR

test <- ndspaScale(test,use ="gmean", method = "nuclei")
}
}
