% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndspaStats.R
\name{ndspaStats}
\alias{ndspaStats}
\title{nDSPA}
\usage{
ndspaStats(
  seobj,
  group.df = NULL,
  group.by = NULL,
  order = NULL,
  segment.tag = "CD45+",
  use.assay = "counts",
  lowexprfilt = FALSE,
  test = c("z.test", "t.test"),
  FDR = TRUE,
  p.value = 0.05,
  fc = 1.5
)
}
\arguments{
\item{seobj}{S4Vector Provide ndspaExperiment object.}

\item{group.df}{data.frame Provide a dataframe object containing group information. Default NULL.In case user manually adds the group information, keep this option NULL.}

\item{group.by}{character Column name common to ndspaExperiment colData and group dataframe. This column will be used to add group information to the colData.}

\item{order}{vector A character vector to define the order in which group1 and group2 should be considered.}

\item{segment.tag}{character Segment tags to be used for performing statistical analysis.}

\item{use.assay}{character Provide the assay name to be used for statistical analysis. Default is "counts" assay (Raw Counts).}

\item{lowexprfilt}{logical To filter out low expressing genes based on min.expr (minimum Signal-to-noise ratio expression; Threshold=1) and min.sm (number of samples required to meet the threshold; Default: 3).
Default: "FALSE"}

\item{test}{character Select which statistical test should be used. Present you can use "z.test" or "t.test".}

\item{FDR}{logical Perform p.value correction. Default "TRUE". (This argument is not yet implemented)}

\item{p.value}{numeric Adjusted p-value cutoff to be used to identify significant Differentially Expressed genes. Default: "0.05"}

\item{fc}{numeric Specify fold change value required for a gene to qualify as Differentially Expressed. Default: "1.5"}
}
\value{
stats.df Returns a data.frame object with lFC and p.adjusted values for significantly Differentially Expressed Genes
}
\description{
Enables to perform group-wise comparison and estimation of statistically significant genes.
}
\details{
The ndspaStats function is let user identify Differentially expresed genes for a given segment and groups.
}
\examples{
\dontrun{
# To plot all QC graphs
dsp.group = read.delim("02-3.dsp_group.sim.tsv", stringsAsFactors = F, header = T)
results <- ndspaStats(test,group.df=dsp.group,group.by="Scan_ID",order=c("NR","R"),segment.tag = "CD45+",test="t.test")


}
}
