% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcfIDsFilter.R
\name{pcfIDsFilter}
\alias{pcfIDsFilter}
\title{nDSPA}
\usage{
pcfIDsFilter(seobj, PCF.min = 0.3, PCF.max = 3)
}
\arguments{
\item{seobj}{S4Vector. Input ndspaExperiment object.}

\item{PCF.min}{numeric. Provide a numeric threshold.}

\item{PCF.max}{numeric. Provide a numeric threshold.}
}
\value{
filtered_scalefactor Vector of Filtered IDs.
}
\description{
Return IDs to be used in ndspaQC.
}
\examples{
\dontrun{
ids <- pcfIDsFilter(test)
}
}
