% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnDSPA.R
\name{readnDSPA}
\alias{readnDSPA}
\title{nDSPA}
\usage{
readnDSPA(x, experiment = c("Protein", "RNA", "CTA"), meta = NULL)
}
\arguments{
\item{x}{character. Specify the file name in .tsv, .csv or .xlsx}

\item{experiment}{character. Specify the experiment type whether it is "Protein", "RNA" or "CTA".}

\item{meta}{This is an experimental argument.
Specify the group file with x,y coordinates (in .csv, .tsv and .xlsx format).}
}
\value{
ne Returns a ndspaExperiment object.
}
\description{
Enables easy loading of the data matrices provided by Nanostring.
}
\details{
The readnDSPA function is built at top of SingleCellExperiment object and is used to load Nanostring spatial omics data.
}
\examples{
\dontrun{
test <- readnDSPA("02-1.dsp_data.raw.sim.tsv", experiment = "RNA",
meta = "02-2.dsp_roi.metadata.sim.tsv")
}
}
