% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshIDsFilter.R
\name{threshIDsFilter}
\alias{threshIDsFilter}
\title{nDSPA}
\usage{
threshIDsFilter(
  seobj,
  FOV = 75,
  bind.min = 0.1,
  bind.max = 2.25,
  min.nuc = 200,
  min.area = 16000
)
}
\arguments{
\item{seobj}{S4Vector. Input ndspaExperiment object.}

\item{FOV}{numeric. Minimum FOV. Default:75 .}

\item{bind.min}{numeric. Minimum Binding Density. Default: 0.1.}

\item{bind.max}{numeric. Maximum Binding Density. Default: 2.25.}

\item{min.nuc}{numeric. Minimum Nucleus Count in AOI. Default:200.}

\item{min.area}{numeric. Minimum area in AOI. Default: 16000.}
}
\value{
character. Vector of Filtered IDs.
}
\description{
Return IDs to be used in ndspaQC based on various Thresholds.
}
\examples{
\dontrun{
ids <- threshIDsFilter(test)
}

}
