% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmean.R
\name{to.numeric}
\alias{to.numeric}
\title{nDSPA}
\usage{
to.numeric(x)
}
\arguments{
\item{x}{numeric. This is a data matrix.}

\item{method}{numeric. The possible values are "log" and "mult". Default Value: "log"}
}
\value{
gm numeric. Returns a gmeans
}
\description{
This function performs basic mean.
}
