% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERCC_Scale_factor.R
\name{ERCC_Scale_factor}
\alias{ERCC_Scale_factor}
\title{nDSPA}
\usage{
ERCC_Scale_factor(df_probe, df_val_all)
}
\arguments{
\item{df_probe}{numeric This is input.}

\item{df_val_all}{numeric Input the datamatrix.}
}
\value{
scalefactor numeric.
}
\description{
This function performs data scaling steps
}
